test(web): move the follow-up urgency view test into web/tests
The module under test is web/ code, so its test belongs in web/tests/
where the web CI already collects it. Relocating it there lets the whole
core-side surface go: the test-all.mjs hook, the update-system.mjs
registration and the core-only-install skip guard all disappear, because
the core runner no longer needs to know the test exists.

Same assertions, same logic under test. followup-view.mjs has no imports
of its own, so the relative import from web/tests/lib needs no resolver
help at all.
