+++ update-system.mjs
@@ -348,7 +348,6 @@ const SYSTEM_PATHS = [
   'test/cover-resolver.test.mjs',
   'test/pipeline-lock.test.mjs',
   'test/profile-photo.test.mjs',
-  'test/followup-view.test.mjs',
   'templates/cv-template.zh-minimal.html',
   'test/zh-minimal-template.test.mjs',
   'test/cv-visual/',
